requirements.txt
python-telegram-bot==13.13
python-dotenv==0.20.0
google-cloud-dialogflow==2.21.0
